requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
Django==4.1.1
django-bootstrap4==22.2
django-filter==22.1
django-model-utils==4.2.0
djangorestframework==3.13.1
importlib-metadata==4.12.0
Markdown==3.4.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pytz==2022.2.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
zipp==3.8.1
